fix(security): add HTML escaping to prevent XSS vulnerabilities
- Create html-utils.ts with escapeHtml and escapeAttr utilities
- Escape user data in charts.ts (bar names, repo names)
- Escape endpoint names in tables.ts
- Escape action labels and onclick handlers in layout.ts
- Escape tool names, endpoint names, and repo names in dashboard.ts
- Export html-utils from components index

Addresses CodeRabbit security review findings
