+++ api/src/templates/components/index.ts
@@ -7,3 +7,4 @@ export * from "./charts";
 export * from "./tables";
 export * from "./metrics";
 export * from "./layout";
+export * from "./html-utils";
